fix(core): support utf8 surrogates
