+++ packages/core/package.json
@@ -50,7 +50,6 @@
     "ajv": "^8.6.0",
     "ajv-errors": "~3.0.0",
     "ajv-formats": "~2.1.0",
-    "blueimp-md5": "2.18.0",
     "es-aggregate-error": "^1.0.7",
     "jsonpath-plus": "7.1.0",
     "lodash": "~4.17.21",
