+++ packages/core/src/spectral.ts
@@ -7,12 +7,10 @@ import { memoize } from 'lodash';
 import { Document, IDocument, IParsedResult, isParsedResult, ParsedDocument } from './document';
 import { DocumentInventory } from './documentInventory';
 import { Runner, RunnerRuntime } from './runner';
-import { IConstructorOpts, IRunOpts, ISpectralDiagnostic, ISpectralFullResult } from './types';
-import { ComputeFingerprintFunc, defaultComputeResultFingerprint } from './utils';
+import type { IConstructorOpts, IRunOpts, ISpectralDiagnostic, ISpectralFullResult } from './types';
+import type { Format, ParserOptions, RulesetDefinition } from './ruleset/index';
 import { Ruleset } from './ruleset/ruleset';
 import { generateDocumentWideResult } from './utils/generateDocumentWideResult';
-import { ParserOptions, RulesetDefinition } from './ruleset/types';
-import { Format } from './ruleset/format';
 import { getDiagnosticSeverity } from './ruleset';
 
 memoize.Cache = WeakMap;
@@ -26,11 +24,7 @@ export class Spectral {
 
   protected readonly runtime: RunnerRuntime;
 
-  private readonly _computeFingerprint: ComputeFingerprintFunc;
-
   constructor(protected readonly opts?: IConstructorOpts) {
-    this._computeFingerprint = memoize(defaultComputeResultFingerprint);
-
     if (opts?.resolver !== void 0) {
       this._resolver = opts.resolver;
     } else {
@@ -81,7 +75,7 @@ export class Spectral {
     }
 
     await runner.run(ruleset);
-    const results = runner.getResults(this._computeFingerprint);
+    const results = runner.getResults();
 
     return {
       resolved: inventory.resolved,
