--- packages/core/src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './prepareResults';
-export * from './sortResults';
